Add make targets for Cilium Mesh and a local smoke-test
In order to test Cilium Mesh functionality locally, one can run
kind-based tests.

There are three ways of preparing the cluster and installing cilium
mesh.  The first one, aka "slow", can be used for an automated
test environment, as this version builds the required images
from sources:

    make kind-cilium-mesh         # Create two kind clusters
    make kind-cilium-mesh-images  # Build & load images into clusters
    make kind-cilium-mesh-install # Install cilium mesh and interconnect tgws

The second way can be used to deploy a cluster from existing images. Instead of
the previous steps, use, e.g.,

    export DOCKER_DEV_ACCOUNT=aspsk DOCKER_REGISTRY=quay.io LOCAL_IMAGE_TAG=cilium-mesh-lab
    make kind-cilium-mesh              # same: create two kind clusters
    make kind-load-cilium-mesh-images  # instead of building images, load them
    make kind-cilium-mesh-install      # Install cilium mesh and interconnect tgws

Now, if you're a developer actively patching code and wanting to test changes
fast (and feeling lucky), then do the following.  First, pull some images from
quay.io (required once):

    docker login quay.io
    docker pull quay.io/isovalent-dev/cilium-ci:main
    docker pull quay.io/isovalent-dev/operator-generic-ci:main
    docker pull quay.io/isovalent-dev/clustermesh-apiserver-ci:main

Now the workflow is as follows:

    make kind-cilium-mesh               # Create two kind-based k8s clusters
    make kind-cilium-mesh-install-fast  # Install cilium
    make kind-image-enterprise-fast     # Build agent and operator and redeploy fast

After this one can do the following in a loop:

    #
    # patch agent/operator/BPF code
    #
    make kind-image-enterprise-fast

Given that one of the two variants was taken, the following command
can be used to test basic connectivity (it will be replaced by a
proper `cilium connectivity test` in future):

    ./contrib/scripts/cilium-mesh-smoke-test.sh

This script is idempotent. If there is a need to clean up test
workloads, then

    ./contrib/scripts/cilium-mesh-smoke-test.sh cleanup

should be used.

For the test itself two gateways (two kind clusters) are attached to the default network:

   GW1
       \
        [docker network "kind"]
       /
   GW2

The cilium-mesh-smoke-test.sh test adds several endpoints, services,
netpolicies, and performs a simple connectivity test between them.

When the testing is done, run

    make kind-cilium-mesh-down

Signed-off-by: Anton Protopopov <aspsk@isovalent.com>