operator/dnsresolver: Fix flakyness in TestManagerPeriodicResolver
TestManagerPeriodicResolver is meant to verify the ability of the
manager to periodically resolve fqdns listed in IsovalentFQDNGroup
custom resources and reconcile the CiliumCIDRGroups accordingly.  The
test relies on a mocked DNS server that returns different IPs for the
same name. Doing so it is possible to verify that the manager eventually
reconciles the CiliumCIDRGroup following an update of the fqdn <-> IPs
mapping.

The previous implementation of the mocked DNS server changed the set of
returned IPs after each query. This was a source of flakyness, since the
returned IPs also have a TTL attached. When the entries expire, the DNS
resolver queries again the mocked DNS server, without coordination with
the main test. This led to a scenario where the mocked DNS server
returned an updated fqdn <-> IPs mapping that was reflected in the
CiliumCIDRGroup, while the test was still expecting the previous one to
appear.

In order to wait for the main goroutine test to verify that the
CiliumCIDRGroup has been reconciled after each unique fqdn <-> IPs
mapping has been received, the mocked DNS server now relies on an atomic
variable to switch to the the next mapping. The atomic variable is
incremented only after the verification step succeeds, to coordinate the
main test goroutine and the mocked DNS server.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>