enterprise/daemon: Add multi-network cell
This commit adds a new cell to implement the `GET /network/attachment`
API call. When queries about the network attachments of a pod, the
current implementation looks at the
`network.v1alpha1.isovalent.com/pod-networks` annotation on the pod. It
then looks up all those networks in via the `IsovalentPodNetwork` CRD
and returns the routes for each configured network. If the annotation on
a pod is not set, we simply return the configuration for the `default`
network, which is then assumed to exist. If the pod object or network
object cannot be found, an error is returned. The API call will be
issued by the CNI plugin, which forwards that error to the container
runtime and eventually kubelet. This means that the request will be
periodically retried while the pod remains in `ContainerCreating`. This
allows the user to create the missing network or fix the faulty pod
annotation, enabling the `GET /network/attachment` call and thereby the
CNI ADD request to eventually succeed.

The API call is only available if the multinetwork cell is enabled
via `enable-multi-network` flag (false by default).

In addition, this commit also adds support for a multi-network
aware auto-direct-node-route feature to Cilium-native multi-homing.
Direct node routes installs a route in the form of `$podCIDR via
$nodeIP` for each remote node in the cluster, thereby allowing pod
traffic to be routed to the connect node on a L2-connected network.

In multi-homing mode, we now need to assume that a node will have
multiple node IP addresses, i.e. at least one per network. Therefore,
this commit adds a new controller which periodically lists the host's
network interfaces and extracts node IP addresses based on the route
listed in the `IsovalentPodNetwork`. If the interface IP matches one of
the network's routes, we then assume it is the canonical IP address of
the node in that network. Future extensions and configurations to this
mechanism are possible, but for now we just support this rather simple
mechanism.

These secondary node IPs are then passed to the `LocalNodeStore`,
which is responsible for announcing them to other nodes in the kvstore
and the CiliumNode CRD. This also has the nice side effect that other
cilium-agents will now add those secondary node IPs into the IPCache
with the `remote-node` identity.

In order for the node kvstore schema to remain compatible to the
upstream one, we currently store the network name in the address type
field. This is a temporary solution, the goal is to eventually upstream
parts of this commit and change the node schema to store the network
name.

Installing the routes is done by watching remote CiliumNodes resources in
the cluster, extracting the remote node's multi-pool pod CIDRs, and then
installing routes in the form of `$podCIDR via $nodeIP` for each node.

Because the CiliumNode CRD only contains the IPAM pool name of a
podCIDR, the code performs a reverse lookup on the `IsovalentPodNetwork`
list to find the corresponding pod network. This works because we
require an IPAM pool to be used by at most one network. Once the network
name of a pod CIDR is known, we can then extract the node's IP in that
network in its `Addresses` section (c.f. previous commit) and install
the routes. To ensure routes are not accidentally deleted by the user,
we also run a reconciliation controller that recomputes all routes for
each node and reinstalls them if necessary.

Ideally, we would integrate this logic into the existing
`auto-direct-node-routes` code in upstream. However, because making the
that code multi-network aware is rather intrusive and would require
kvstore schema changes, we instead let the logic live in the
multinetwork cell for now. One shortcoming of this (besides code
duplication) is that the multinetwork version only subscribes to
CiliumNode updates, which means clustermesh is not supported.

Co-authored-by: Tobias Klauser <tobias@isovalent.com>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>
Signed-off-by: Tobias Klauser <tobias@isovalent.com>