enterprise/cilium-cni: Add mindful CNI plugin
This commit overwrites the Cilium CNI plugin build with an
enterprise version, allowing the Cilium Enterprise codebase to modify
the logic of the binary. This pattern is very similar to how we also
overwrite the cilium-agent or cilium-operator binary in the enterprise
codebase.

This commit does not yet contain any enterprise-specific logic. This
will follow in a subsequent commit.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>