enterprise/cilium-cni: Add multinetwork support
This commit extends the Cilium CNI plugin with the ability to attach
multiple interfaces to a container during a CNI ADD invocation.  This is
done by first determining the configured networks on the pod via the
`GET /network/attachment` API call and then executing the main CNI ADD
logic for each network attachment (aka. network interface):

1. We allocate an IP using multi-pool IPAM for the interface. While not
enforced by the CNI plugin at the moment, we do not support multiple
networks to share the same IPAM pool, i.e. for a given IPAM pool there
will only ever be one network it can be used in. This is implemented by
returing the selected IPAM pool in `EndpointConfiguration.IPAMPool`.

2. We prepare an endpoint creation request with the container interface
field populated the endpoint model. Here, we distinguish between the
"primary" endpoint (which is identical to the endpoint which would be
created if a pod only attached to a single interface) and "secondary"
endpoints. Disable "legacy endpoint identifiers", meaning
their CEP name will be suffixed with the container interface name.
Another difference exists between endpoints attached to the "default"
network and endpoints attached to a non-default network. Endpoints on
the non-default network contain the network attachment name as an
additional label, thereby allowing each interface to have it's own Cilium
security identity for policy selection.

3. When setting up each network interface, we install the additional
routes listed in the `IsovalentPodNetwork` of the attachment. This
will make sure that traffic destined to a secondary network is routed
via the secondary interface.

4. An endpoint is then created for each network attachment. By having
multiple endpoints per pod, other Cilium components do not need to be
made aware of multi-homing. The necessary refactors in the endpoint
manager were already done in previous pull requests

The CNI DEL already gained support for multi-homed pods in an earlier
commit where we added support for endpoint batch delete in the endpoint
deletion queue. By issuing a batch delete of all endpoints with a
particular container ID, we delete all the endpoints created by an
earlier CNI ADD call.

Co-authored-by: Tobias Klauser <tobias@isovalent.com>
Signed-off-by: Tobias Klauser <tobias@isovalent.com>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>