egressgwha: operator: manager: add support for AZ affinity
this commit adds support to the egressgwha operator manager to AZ
affinity.

With this change the operator will calculate, in addition to the usual
activeGatewayIPs and healthyGatewayIPs status lists, a new
activeGatewayIPsByAZ status for the active gateways for any availability
zone.

Here's an example of the new status:

    status:
      groupStatuses:
      - activeGatewayIPs:
        - 192.168.80.4
        - 192.168.80.5
        activeGatewayIPsByAZ:
          us-east-1a:
          - 192.168.80.4
          us-east-1b:
          - 192.168.80.5
        healthyGatewayIPs:
        - 192.168.80.4
        - 192.168.80.5
      observedGeneration: 2

Support for the agent manager will be added in a subsequent commit.

Signed-off-by: Gilberto Bertin <jibi@cilium.io>