cli: Fix incollect sysdump of IsovalentSRv6LocatorPool/SIDManager
We mistakenly reintroduced the bug fixed in
fd14a361a2daf765c91b90c53524b783de679fb7 that the
IsovalentSRv6LocatorPool and IsovalentSRv6SIDManager is not corrected
properly because of the name of the resources are not specified in
prural name when we port SRv6 Mananger to main-ce in #2020. Fix it
again.

Fixes: #2020

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>