enterprise/daemon: add enterprise featurelist to daemon cells.
Initially, this will be committed as an empty cell Group and added to the enterprise Hive.
This is to provide just the scafolding for how legacy features will be added and layered on top of the OSS core of Cilium CE.
As well, subsequent commits will update CI to use the "allow-all" configuration for CI workflows as to avoid feature checking errors during workflows that use main branch commits (i.e. helm-defaults).

Note: The enterprise/pkg/featurelist package should be for features that cannot be declared alongside their Enterprise Cells - such as existing OSS features that need to be feature gated in CE.

Signed-off-by: Tom Hadlaw <tom.hadlaw@isovalent.com>