enterprise/pkg/features: add features package.
enterprise/pkg/features implements a FeatureWithConfigT[T] cell type that:

* Add feature gate flags using Config[T].
* Declares and registers a new feature specification.
* Checks if feature can be enabled given current feature flag configuration.

Currently, this will check if a feature (which is not allowed by default),
has the appropriate readiness feature gate enabled to be toggled on.

Features can be registered in a Hive using the features.FeatureWithConfigT[T](...) type.

For example:

```
// ...
features.FeatureWithConfigT[exampleFeatureConfig](features.Spec{
	ID:          "example",
	Name:        "Example Feature",
	Description: "This here doohickey is an example feature.",
	Default:     false,
	Since:       semver.Version{Major: 1, Minor: 14},
	Stage:       features.Alpha,
}),
// ...
```

Also provides optional WithIsEnabledFn featureOpt[T] func which instruments the
FeatureWithConfigT to use custom override function to determine whether a feature is enabled.
This is intended as an escape hatch to cover almost all cases where a feature is not necessarily
toggled by a single XXXEnabled bool value (ex. ip dual-stack).

Signed-off-by: Tom Hadlaw <tom.hadlaw@isovalent.com>