hubble-flow-export: Test flow aggregation on export
This change required plumbing the clock to the underlying cache from the
aggregator into every intermediate object, but allows controlling the
clock so we can test the TTL on aggregation behaves correctly within the export
plugin.

This also require updating unit tests in the aggregation plugin to use a
fakeclock which also meant we could remove sleeps, and replace them with
advancing the fake clock, which is a nice improvement.

Signed-off-by: Chance Zibolski <chance.zibolski@gmail.com>