mixed routing: introduce GHA E2E test workflow
Introduce a new GitHub actions workflow which validates the correct
operation of the mixed routing mode for clustermesh functionality, and
more in general of the fallback routing mode configuration. The overall
goal is to assert that (a) enabling the fallback routing mode does not
disrupt long running connections and (b) each node uses the correct
routing mode towards every other node, depending on the primary and
fallback routing mode configured for each cluster.

The test validates a set of different configurations, including
combinations of KPR enabled and disabled, endpoint routes enabled and
disabled, tunneling protocols, encryption modes (WireGuard only, as
IPSec is not supported at the moment), and clustermesh modes (i.e.,
clustermesh-apiserver, kvstoremesh and external kvstore). Additionally,
we test the possible routing mode configurations, i.e., native-native,
tunnel-tunnel, and native-tunnel. The first two to assert that we don't
break connectivity for clusters already configured with the same routing
mode, while the last tests an actual mixed routing mode configuration.
We decided to adopt this matrix based approach, rather than creating an
higher number of clusters in a single run to limit the total duration
of the tests (i.e., not run multiple connectivity suites sequentially)
and avoid overloading the GH runners with an excessive number of kind
workers.

More specifically, the test initially deploys two kind clusters, installs
Cilium and, when the routing modes match, meshes the two clusters
together. At this point, it starts the pods to check for long running
connections disruption. With all of that in place, it then enables the
mixed routing mode support in both clusters, and asserts that existing
connections haven't been disrupted. Finally, it meshes the two clusters
together when they are configured with different routing modes, and runs
the full Cilium CLI connectivity suite tests. During the execution of
the different tests, the suite will also assert that the correct routing
mode is selected towards both local and remote nodes.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>