enterprise/cilium-cli: mention version update in RELEASE.md
This was missed during the v0.15.20-cee.1 release in commit e5eb927d7493
("enterprise/cilium-cli: prepare for v0.15.20-cee.1 release"). Update
the version and mention in in the release instructions so we hopefully
don't forget to bump the version during future releases.

Also link the GH action output for the tag pushes which makes it easier
to find the relevant run.

Signed-off-by: Tobias Klauser <tobias@isovalent.com>