sidmanager: Add ErrAllocatorNotFound error
Add ErrAllocatorNotFound error which can be handled with errors.Is().
This is useful to handle the case that ManageSID() cannot be performed
because of the missing allocator. Sometimes this error can be ignored
for example in the case when IsovalentSRv6LocatorPool is already
created, but the allocator is not yet created locally (which happens
when cilium-operator is not running).

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>