srv6: Rename IsovalentVRF.spec.podSelector to endpointSelector
Rename IsovalentVRF.spec.podSelector field to endpointSelector. We don't
need to change any agent logic because the agent already selects
endpoints instead of Pods. This is because SRv6 Manager is created when
there is a 1to1 mapping between an endpoint and a Pod (after
multinetwork, it is not the case). So, this change is about aligning the
field name with the actual semantics.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>