ent/cli: introduce mixed routing mode connectivity tests
Introduce a new connectivity test to cover the correct operation of the
mixed routing mode for clustermesh functionality, and more in general of
the fallback routing mode configuration. The overall goal is to assert
that each node uses the correct routing mode towards every other node,
be it direct routing or tunneling.

To this end, we leverage an approach similar to that used for encryption
tests: we first start a tcpdump capture on the host-netns pods created by
the connectivity suite, configuring the appropriate filter (i.e., checking
that no vxlan/geneve packet is exchanged between nodes configured in
direct routing mode, as well as no packet with source/destination belonging
to the PodCIDR is observed for nodes operating in tunnel mode). Finally,
we assert that no packet matching the given filters was captured. To
prevent unexpectedly breaking the filters, we additionally run the
sniffers in sanity mode, hence asserting that we observe the expected
packets when the filter matches the configured routing mode.

It is worth mentioning that the test itself does not generate any traffic,
but relies on the background traffic exchanged inside and between clusters,
as well as that generated by the rest of the connectivity suite. Hence,
it is expected that false positives may occur due to the sanity checks
when run in combination with a reduced set of tests, or in a large,
underused cluster. Yet, this is deemed acceptable for the moment, given
that the test already belongs to the unsafe suite (which essentially means
we'll run it in controlled environments only), and the benefits of the
sanity mode largely overweight possible seldom failures (which are not
expected to occur in CI at all).

One minor limitation is that the setup of the sniffers is always
performed, regardless of whether the test has been manually skipped by
the user. This stems from the difficulty of checking in advance whether
a given test/scenario combination would be executed or not. Yet, this
is not deemed to be an issue given that both setup and validation are
performed only if the unsafe suite is enabled.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>