gha: bump kindest/node version in overlapping podcidr workflow
Bump the kindest/node image version used by the ClusterMesh Overlapping
PodCIDR to v1.29.0-rc.1 to match the one used by other workflows.
Additionally configure renovate to source the version from the
cilium/kindest-node quay repository instead of kindest/node, to align
with the rest of workflows, and match the image specified in the template.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>