egressgatewayha: expose EGW IPs to BGP Control Plane
[ upstream commit ed2c970dd942e2c23960bb398b348e6981c493a9]

Implements a new EGW Manager method `AdvertisedEgressIPs` which returns
a list of egress IPs that should be advertised via BGP based on provided
EGW policy label selector.

To allow that, the internal `PolicyConfig` struct in egressgatewayha
is extended to store policy lables.

Uses the `BGPCPSignaler` in the EGW Manager to signal BGP Control Plane
when reconciliation is needed.

Signed-off-by: Rastislav Szabo <rastislav.szabo@isovalent.com>
Signed-off-by: harsimran pabla <hpabla@isovalent.com>