mixed routing: extend validation to include more incompatibilities
Let's extend the validation logic for mixed routing mode to also bail
out loudly in case of other known incompatibilities, to prevent users
from running unsupported and/or untested feature combinations. Most
notably, we forbid encryption (the WireGuard limitation is planned to
be lifted in a future PR), Egress Gateway (it is not compatible with
clustermesh in general, but mixed routing mode may break it also within
a single cluster), Overlapping PodCIDR (already requires tunneling)
and a bunch of additional, rarer, configurations.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>