mixed routing: lift the WireGuard incompatibility
Now that mixed routing mode for clustermesh works in combination
with WireGuard, let's not complain anymore if both are enabled at
the same time. However, let's still forbid turning on Wireguard
node-to-node encryption and strict mode, as beta features, and not
tested at all in combination with mixed routing mode.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>