egressgwha: operator: manager: select active GWs per AZ with random probability
Currently, the active GWs selection logic for AZ affinity chooses the
local active GWs from the first in order of the node name, and if it's
not enough, then selects non-local GWs from AZ next to the target AZ
passed to the function, and picks the required number of the GWs.
(The AZ slice is sorted in lexicographical ordering)

- non local gateways from az-1 will be selected from az-2, az-3,..
- non local gateways from az-2 will be selected from az-3, az-4,..

The current approach is too simplistic and has an issue in a case
where AZ have vastly different number of gateway nodes. This always
more heavily loads nodes early in the per AZ list.

This commit improves the active GWs selection logic to fix the issue
by selecting the active GWs with random probability.

Signed-off-by: Yusuke Suzuki <yusuke.suzuki@isovalent.com>