ent/cli: add extra test to avoid mixed routing false positives
The mixed routing tests perform a sanity check to ensure that we observe
packets matching the constructed filters, and guarantee that we don't
silently break them. However, they rely on some traffic being exchanged
between pods running on different nodes of the clusters, which may not
always be the case in CI environments. One particular case being when
WireGuard encryption is enabled, because node to pod traffic (i.e.,
health checks) is not observed as not encrypted in native routing
clusters (it would require node-to-node encryption to be enabled).

To avoid false positives, let's add an additional test which generates
traffic from one pod in each cluster towards all other nodes. To this
end, we target the Cilium HealthIP, as belonging to the PodCIDR of
each node. One limitation of this check (to avoid needing to deploy
additional pods) is that it requires health checking to be enabled.
However, this is sufficient for our current purpose, given that it
is always the case in CI environments.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>