bpf: encrypt overlay traffic
This commit introduces the ability to encrypt overlay traffic before it
leaves the host.

The 'cil_to_netdev' function is updated to sniff into overlay packets
(only VXLAN supported for now) and determine if an encryption mark is
set in the overlay's header.

If it is, a new function in encrypt.h will set the appropriate packet
mark on the skb and redirect the packet to the ingress of the interface
it was egressing on.

When the packet is seen on the ingress side of the device it will be
submitted to the XFRM hooks in the output routing path and the XFRM
subsystem will encrypt the packet.

Subsequent changes to the IPSec control plane to create the appropriate
states and policies to support this are required.

Signed-off-by: ldelossa <louis.delos@isovalent.com>