gha: rework wait for policy enforcement in Overlapping PodCIDR test
ed817e6f6777 ("Remove additional endpoint status options.") dropped the
opt-in support for propagating policy enforcement information to the
CiliumEndpoint CRD, which we relied upon in this workflow. Hence, let's
rework the logic to retrieve the same information using cilium-dbg, to
ensure that we continue with the rest of the test only when policies are
confirmed to be in place.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>