ent/cli: tune mixed routing tcpdump filter to avoid false positive
I've been recently witnessing a flake caused by the mixed routing mode
assertion test reporting the detection of an unexpected packet forwarded
natively:

  IP6 fd00:10:244:2::d7d9.8080 > fd00:10:242:1::edd3.44301: Flags [F.]

where the source IP address is that of the echo-other-node pod, and the
destination address is one of the Cilium Ingress IPs. Upon further
investigation, the sequence of events appears to be the following:

1. The connectivity suite executes one of the pod-to-ingress-service tests;
2. After 5 seconds, the server attempts to close the connection (as not
   closed by Envoy to hopefully reuse it), and at that point sends a FIN
   packet to the client (i.e., Ingress IP). Yet, the connectivity suite
   already configured a different ingress policy, causing the packet to
   be dropped. The FIN packet is then retransmitted multiple times.
3. Eventually, the server may succeed sending an ACK, followed by a FIN
   packet when no policies blocking them are currently applied. At that
   point, the first ACK causes the disposal of the socket (as the other
   direction was already closed), and the subsequent FIN doesn't match
   a valid socket anymore. Hence, it gets routed according to the default
   route (as the IP is not assigned to any local interface, and it doesn't
   match any more specific route), and it is flagged by the mixed routing
   mode test because it goes through the incorrect interface.

The same situation can be easily obtained curling the Ingress IP
(selecting a random port not corresponding to any active socket), and
observing that the SYN packet gets incorrectly forwarded, rather
than being swallowed, optionally emitting an RST packet.

As a short term solution, let's amend the tcpdump filter used by the
mixed routing mode tests to ignore packets having as destination one of
the local Ingress IPs. This prevents the flake, while not reducing the
overall coverage at the same time, because we would still detect every
legitimate failure on the source node side.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>