.github: use single quote for cilium-config action.
It seems like there is an issue with using double quotes to pass this array.
As well in order to test subsequent changes this needs to be merged into main-ce first.
This should fix issue where the array being passed was being mangled and only the first feature gate value was passed (ex. [AllAlphaFeatures]).

Signed-off-by: Tom Hadlaw <tom.hadlaw@isovalent.com>