egressgwha: operator: manager: improve active GWs selection for non-AZ affinity
Currently, we calculate the active GWs twice when active GW restarts
(down and up event). Supposing node1 goes down while the active GWs
are node1 and node2, we select node2 and node3 as the active GWs.
And then we select node1 and node2 again when node1 is up.
This results in sessions to break twice. Ideally, once failover has
happened it should stay there till active node goes down.

Instead of re-calculating the active gateways, we discard all unhealthy
gateways and non gataway nodes from the active gateways, and then
back-fill the currently selected group of active gateways, until
maxGateways is reached.

Note that we ignore the previous status if IEGP is updated and the
operator hasn't seen its latest spec. Otherwise we coulde feed a wrong
status to the calculation algorithm.

This commit applies the improvement to the non AZ affinity case.

fixes: #1722

Signed-off-by: Yusuke Suzuki <yusuke.suzuki@isovalent.com>