egressgwha: operator: manager: improve active GWs selection for AZ affinity case
This commit applies the improvement of the active GWs selection for the AZ affinity case.

We split the current active gateways into the local and non-local, and apply the following
logic to the local gataways first. If it's not enough, apply the same logic to the non-local
gateways until maxGateways is reached.

- discard all unhealthy gateways and non gataway nodes from the active gateways, and then
- back-fill the currently selected group of active gateways, until maxGateways is reached.

For example, we have 5 gateway nodes in 2 AZs.
Let's say we select 2 gateways for az-1, maxGW=2 and azAffinity=localPriority

{ az-1: gw-1 gw-2 gw-3, az-2: gw-4 gw-5 }

We select 2 local gateways from the healthy local gateways(gw-1 gw-2 gw-3)

{ az-1: gw-1, gw-2 }

If gw-2 goes down, we discard it, select a healthy local gateway(gw-3 in this case)
and back-fill the currently selected group of local active gateways

{ az-1: gw-1, gw-3 }

If gw-2 is up again, we don't recalculate the active gateways because the condition is already met.

{ az-1: gw-1, gw-3 }

If gw-2 and gw-3 go down, we discard gw-3 and select a healthy non-local gateway
(gw-4 in this case) because there's no healthy local gateway available other than gw-1.

{ az-1: gw-1, gw-4 }

If gw-4 goes down, we discard gw-4 and select a healthy non-local gateways(gw-5 in this case)

{ az-1: gw-1, gw-5 }

If gw-4 is up again, we don't recalculate the active gateways because the condition is already met.

{ az-1: gw-1, gw-5 }

If gw-2 is up, we select gw-2 and discard gw-5. Because a local gateway is available now.

{ az-1: gw-1, gw-2 }

Signed-off-by: Yusuke Suzuki <yusuke.suzuki@isovalent.com>