sysdump: no need for tetragon pod logs, configmap
With
https://github.com/cilium/cilium-cli/commit/59ffc681cab678b3db6c73d86029db5a4048c2c9
merged in cilium-cli, tetragon pod logs and the tetragon config map are
gathered in OSS side, so no need to have EE-specific tasks for them.

The duplicated tasks are executed twice, but they result in the same
file. Because the tasks end up using os.WriteFile() [1], the file is
opened with O_TRUNC so if the tasks do not run in parallel we get only
one log file with the appropriate data. Haven't tried it, but I think
there is a potential of corruption of the tasks run in parallel.

In any case, there is no need to do duplicate work so we remove the
extra tasks.

This was tested manually by checking that the sysdump included the pod
logs and the config map.

[1] https://cs.opensource.google/go/go/+/refs/tags/go1.22.1:src/os/file.go;l=814

Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>