sysdump: remove tetragon code for sysdump
With
https://github.com/cilium/cilium-cli/commit/1632e1b0d97dd3ce61a0c153bf66e79e9a46dc4f
being merged in the OSS cilium-cli, we use ListUnstructured to get the
tetragon CRs (tracingpolicies and podinfo) so there is no need to import
the tetragon types anymore. Remove them.

This effectively reverts
https://github.com/isovalent/cilium/commit/b968f2a0f4f02c495d71edd61b343d38154c8c03, which introduced the tetragon directory.

This was tested manually by checking that the tracingpolicies in the
sysdump included the file: section of a tracing policy.

Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>