mixed routing: rework ipset entries filtering based on routing mode
Normally, Cilium inserts an ipset entry for each NodeInternalIP address
when configured in native routing mode if iptables masquerading is
enabled, to skip masquerading for pod to node traffic. Differently,
masquerading must be performed when running in tunnel mode, otherwise
packets may be dropped, and return traffic would not work.

This commit reworks the previous implementation, removing the wrapper
around the IPSetManager used to filter out ipset upsertion operations
based on the exclusion list configured when new node upsertions and
deletions are observed. Differently, we simply inject a filter function,
which takes a node object as parameter, into Hive, leveraged by the
OSS logic to directly filter out the upsertion operation. That way,
we can directly check the routing mode annotation, avoiding additional
indirections which are potentially fragile (and likely to cause build
conflicts upon OSS changes). We couldn't adopt this approach earlier
because the previous IPSetManager implementation didn't remove possible
stale entries, which could have caused issues if the routing mode was
changed when the Cilium agent was down.

Similarly to before, this logic works under the assumption that the
fallback routing mode is tunnel (i.e., the only supported configuration
at the moment). Hence, it filters out the ipset entries associated
with nodes towards which the preferred routing mode is tunnel.
To support native routing fallback, instead, we should also enable
the ipset manager logic in that case (as otherwise disabled).

Fixes: ca0ef85546b4 ("mixed routing: filter ipset entries creation based on routing mode")
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>