fqdnha: don't set HA FQDN port twice
This commit removes the unnecessary config change in the proxy config
to set the FQDN HA port to `10001`.

This shouldn't be necessary as the proxy port is already set when
calling `L7Proxy.SetProxyPort(...)` with the same port in
`enterprise/pkg/fqdnha/remoteproxy/proxy.go`.

This is executed in the config phase of the Hive initialization - hence
before bootstrapping the FQDN proxy in the Cilium Agent that is asking the
Proxy whether there's already a FQDN port set.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>