srv6: Decouple SIDStructure from SID and Locator
In the upcoming new SID allocation scheme, the locator prefix and the
actual structure doesn't match. For example, in the current
implementation, we could create a locator like:

Prefix: fd00:1234:5678::/48
Structure: [32, 16, 16, 0]

Which the locator prefix length matches the structure. However, in the
new allocation scheme, we want to create a locator like:

Prefix: fd00:1234:5678:9abc::/64
Structure: [32, 16, 32, 0]

Which the locator prefix length doesn't match the structure. Once we
allow doing this, associating structure with Locator or SID doesn't make
any sense. Thus, decouple structure from SID and Locator. Now the SID
and Locator types are a wrapper of netip.Addr/Prefix with some useful
methods for SRv6 use cases.

This commit is made in the following steps:

1. Remove the structure field from Locator and SID types. Most of the
   methods are left as is, but modified to take a structure as an
   argument.
2. Fix all compile error introduced by 1.
3. Make tests work. The tricky part was equality checks for SID and
   Locator (e.g., sid1 == sid2) used to compare structure as well, but
   after removing the structure field, we need to compare the structure
   separately (e.g., sid1 == sid2 && structure1 == structure2).
   require.Equal or assert.Equal used in tests have the same issue. Fix
   them.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>